+++ src/main/java/com/getindata/connectors/http/internal/SinkHttpClientBuilder.java
@@ -6,7 +6,8 @@ import java.util.Properties;
 import org.apache.flink.annotation.PublicEvolving;
 
 import com.getindata.connectors.http.HttpPostRequestCallback;
-import com.getindata.connectors.http.internal.sink.HttpSinkRequestEntry;
+import com.getindata.connectors.http.internal.sink.httpclient.HttpRequest;
+import com.getindata.connectors.http.internal.sink.httpclient.RequestSubmitterFactory;
 
 /**
  * Builder building {@link SinkHttpClient}.
@@ -14,11 +15,13 @@ import com.getindata.connectors.http.internal.sink.HttpSinkRequestEntry;
 @PublicEvolving
 public interface SinkHttpClientBuilder extends Serializable {
 
-    // TODO Consider moving HttpPostRequestCallback and HeaderPreprocessor to be a
+    // TODO Consider moving HttpPostRequestCallback and HeaderPreprocessor, RequestSubmitter to be a
     //  SinkHttpClientBuilder fields. This method is getting more and more arguments.
     SinkHttpClient build(
         Properties properties,
-        HttpPostRequestCallback<HttpSinkRequestEntry> httpPostRequestCallback,
-        HeaderPreprocessor headerPreprocessor
+        HttpPostRequestCallback<HttpRequest> httpPostRequestCallback,
+        HeaderPreprocessor headerPreprocessor,
+        RequestSubmitterFactory requestSubmitterFactory
+
     );
 }