+++ src/main/java/com/getindata/connectors/http/internal/config/SinkRequestSubmitMode.java
@@ -1,0 +1,17 @@
+package com.getindata.connectors.http.internal.config;
+
+public enum SinkRequestSubmitMode {
+
+    SINGLE("single"),
+    BATCH("batch");
+
+    private final String mode;
+
+    SinkRequestSubmitMode(String mode) {
+        this.mode = mode;
+    }
+
+    public String getMode() {
+        return mode;
+    }
+}