+++ src/main/java/com/getindata/connectors/http/internal/sink/HttpSinkInternal.java
@@ -16,6 +16,12 @@ import com.getindata.connectors.http.HttpPostRequestCallback;
 import com.getindata.connectors.http.SchemaLifecycleAwareElementConverter;
 import com.getindata.connectors.http.internal.HeaderPreprocessor;
 import com.getindata.connectors.http.internal.SinkHttpClientBuilder;
+import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants;
+import com.getindata.connectors.http.internal.config.SinkRequestSubmitMode;
+import com.getindata.connectors.http.internal.sink.httpclient.BatchRequestSubmitterFactory;
+import com.getindata.connectors.http.internal.sink.httpclient.HttpRequest;
+import com.getindata.connectors.http.internal.sink.httpclient.PerRequestRequestSubmitterFactory;
+import com.getindata.connectors.http.internal.sink.httpclient.RequestSubmitterFactory;
 
 /**
  * An internal implementation of HTTP Sink that performs async requests against a specified HTTP
@@ -59,7 +65,7 @@ public class HttpSinkInternal<InputT> extends AsyncSinkBase<InputT, HttpSinkRequ
     // makes it possible to serialize `HttpSink`
     private final SinkHttpClientBuilder sinkHttpClientBuilder;
 
-    private final HttpPostRequestCallback<HttpSinkRequestEntry> httpPostRequestCallback;
+    private final HttpPostRequestCallback<HttpRequest> httpPostRequestCallback;
 
     private final HeaderPreprocessor headerPreprocessor;
 
@@ -74,7 +80,7 @@ public class HttpSinkInternal<InputT> extends AsyncSinkBase<InputT, HttpSinkRequ
         long maxTimeInBufferMS,
         long maxRecordSizeInBytes,
         String endpointUrl,
-        HttpPostRequestCallback<HttpSinkRequestEntry> httpPostRequestCallback,
+        HttpPostRequestCallback<HttpRequest> httpPostRequestCallback,
         HeaderPreprocessor headerPreprocessor,
         SinkHttpClientBuilder sinkHttpClientBuilder,
         Properties properties) {
@@ -127,7 +133,12 @@ public class HttpSinkInternal<InputT> extends AsyncSinkBase<InputT, HttpSinkRequ
             getMaxTimeInBufferMS(),
             getMaxRecordSizeInBytes(),
             endpointUrl,
-            sinkHttpClientBuilder.build(properties, httpPostRequestCallback, headerPreprocessor),
+            sinkHttpClientBuilder.build(
+                properties,
+                httpPostRequestCallback,
+                headerPreprocessor,
+                getRequestSubmitterFactory()
+            ),
             Collections.emptyList(),
             properties
         );
@@ -149,7 +160,12 @@ public class HttpSinkInternal<InputT> extends AsyncSinkBase<InputT, HttpSinkRequ
             getMaxTimeInBufferMS(),
             getMaxRecordSizeInBytes(),
             endpointUrl,
-            sinkHttpClientBuilder.build(properties, httpPostRequestCallback, headerPreprocessor),
+            sinkHttpClientBuilder.build(
+                properties,
+                httpPostRequestCallback,
+                headerPreprocessor,
+                getRequestSubmitterFactory()
+            ),
             recoveredState,
             properties
         );
@@ -157,7 +173,16 @@ public class HttpSinkInternal<InputT> extends AsyncSinkBase<InputT, HttpSinkRequ
 
     @Override
     public SimpleVersionedSerializer<BufferedRequestState<HttpSinkRequestEntry>>
-        getWriterStateSerializer() {
+            getWriterStateSerializer() {
         return new HttpSinkWriterStateSerializer();
+    }
+
+    private RequestSubmitterFactory getRequestSubmitterFactory() {
+
+        if (SinkRequestSubmitMode.SINGLE.getMode().equalsIgnoreCase(
+            properties.getProperty(HttpConnectorConfigConstants.SINK_HTTP_REQUEST_MODE))) {
+            return new PerRequestRequestSubmitterFactory();
+        }
+        return new BatchRequestSubmitterFactory(getMaxBatchSize());
     }
 }