+++ src/test/java/com/getindata/connectors/http/internal/HttpsConnectionTestBase.java
@@ -4,6 +4,9 @@ import java.util.Properties;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 
+import com.getindata.connectors.http.HttpPostRequestCallback;
+import com.getindata.connectors.http.internal.sink.httpclient.HttpRequest;
+import com.getindata.connectors.http.internal.table.sink.Slf4jHttpPostRequestCallback;
 import com.getindata.connectors.http.internal.utils.HttpHeaderUtils;
 
 public abstract class HttpsConnectionTestBase {
@@ -27,6 +30,9 @@ public abstract class HttpsConnectionTestBase {
     protected Properties properties;
 
     protected HeaderPreprocessor headerPreprocessor;
+
+    protected HttpPostRequestCallback<HttpRequest> postRequestCallback =
+        new Slf4jHttpPostRequestCallback();
 
     public void setUp() {
         this.properties = new Properties();