+++ src/test/java/com/getindata/connectors/http/internal/sink/HttpSinkBuilderTest.java
@@ -21,11 +21,13 @@ public class HttpSinkBuilderTest {
     public void testEmptyUrl() {
         assertThrows(
             IllegalArgumentException.class,
-            () -> HttpSink.<String>builder()
-                .setElementConverter(ELEMENT_CONVERTER)
+            () -> HttpSink.<String>builder().setElementConverter(ELEMENT_CONVERTER)
                 .setSinkHttpClientBuilder(
-                    (properties, httpPostRequestCallback, headerPreprocessor)
-                        -> new MockHttpClient())
+                    (
+                        properties,
+                        httpPostRequestCallback,
+                        headerPreprocessor,
+                        requestSubmitterFactory) -> new MockHttpClient())
                 .setEndpointUrl("")
                 .build()
         );
@@ -38,8 +40,11 @@ public class HttpSinkBuilderTest {
             () -> HttpSink.<String>builder()
                 .setElementConverter(ELEMENT_CONVERTER)
                 .setSinkHttpClientBuilder(
-                    (properties, httpPostRequestCallback, headerPreprocessor)
-                        -> new MockHttpClient())
+                    (
+                        properties,
+                        httpPostRequestCallback,
+                        headerPreprocessor,
+                        requestSubmitterFactory) -> new MockHttpClient())
                 .build()
         );
     }