+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## [Unreleased]
 
+### Added
+-   Add support for batch request submission in HTTP sink. The mode can be changed by setting
+    `gid.connector.http.sink.writer.request.mode` with value `single` or `batch`. The default value
+    is `batch` bode which is breaking change comparing to previous versions. Additionally,
+    `gid.connector.http.sink.request.batch.size` option can be used to set batch size. By default,
+    batch size is 500 which is same as default value of HttpSink `maxBatchSize` parameter. 
+
+### Changed
+-   Changed API for public HttpSink builder. The `setHttpPostRequestCallback` expects a `PostRequestCallback`
+    of generic type [HttpRequest](src/main/java/com/getindata/connectors/http/internal/sink/httpclient/HttpRequest.java)
+    instead `HttpSinkRequestEntry`.
+-   Changed HTTP sink request and response processing thread pool sizes from 16 to 1.
+
 ## [0.9.0] - 2023-02-10
 
 -   Add support for Flink 1.16.